Cleanup, organizing (MCChatListener)
Removed all sysouts
Split MCChatListener into utils, private and custom (and the existing one)
Using fixed version of Discord4J (fixes issues with local testing)
