+++ src/main/java/buttondevteam/discordplugin/DiscordPlayer.java
@@ -1,6 +1,6 @@
 package buttondevteam.discordplugin;
 
-import buttondevteam.discordplugin.listeners.MCChatListener;
+import buttondevteam.discordplugin.mcchat.MCChatPrivate;
 import buttondevteam.lib.player.ChromaGamerBase;
 import buttondevteam.lib.player.UserClass;
 
@@ -20,9 +20,9 @@ public class DiscordPlayer extends ChromaGamerBase {
 
 	/**
 	 * Returns true if player has the private Minecraft chat enabled. For setting the value, see
-	 * {@link MCChatListener#privateMCChat(sx.blah.discord.handle.obj.IChannel, boolean, sx.blah.discord.handle.obj.IUser, DiscordPlayer)}
+	 * {@link MCChatPrivate#privateMCChat(sx.blah.discord.handle.obj.IChannel, boolean, sx.blah.discord.handle.obj.IUser, DiscordPlayer)}
 	 */
 	public boolean isMinecraftChatEnabled() {
-		return MCChatListener.isMinecraftChatEnabled(this);
+		return MCChatPrivate.isMinecraftChatEnabled(this);
 	}
 }
