+++ src/main/java/buttondevteam/discordplugin/commands/MCChatCommand.java
@@ -2,7 +2,7 @@ package buttondevteam.discordplugin.commands;
 
 import buttondevteam.discordplugin.DiscordPlayer;
 import buttondevteam.discordplugin.DiscordPlugin;
-import buttondevteam.discordplugin.listeners.MCChatListener;
+import buttondevteam.discordplugin.mcchat.MCChatPrivate;
 import buttondevteam.lib.TBMCCoreAPI;
 import sx.blah.discord.handle.obj.IMessage;
 
@@ -22,7 +22,7 @@ public class MCChatCommand extends DiscordCommandBase {
 		}
 		try (final DiscordPlayer user = DiscordPlayer.getUser(message.getAuthor().getStringID(), DiscordPlayer.class)) {
 			boolean mcchat = !user.isMinecraftChatEnabled();
-			MCChatListener.privateMCChat(message.getChannel(), mcchat, message.getAuthor(), user);
+			MCChatPrivate.privateMCChat(message.getChannel(), mcchat, message.getAuthor(), user);
 			DiscordPlugin.sendMessageToChannel(message.getChannel(),
 					"Minecraft chat " + (mcchat //
                             ? "enabled. Use '/mcchat' again to turn it off." //
